+++ hydra-plutus/src/Hydra/Contract/Util.hs
@@ -8,10 +8,11 @@ import Plutus.V2.Ledger.Api (
   CurrencySymbol,
   TokenName (..),
   TxInfo (TxInfo, txInfoMint),
-  Value (getValue),
+  Value (getValue), toBuiltinData,
  )
 import qualified PlutusTx.AssocMap as Map
 import PlutusTx.Prelude
+import PlutusTx.Builtins (serialiseData)
 
 hydraHeadV1 :: BuiltinByteString
 hydraHeadV1 = "HydraHeadV1"
@@ -47,3 +48,15 @@ mustNotMintOrBurn TxInfo{txInfoMint} =
     isZero txInfoMint
 {-# INLINEABLE mustNotMintOrBurn #-}
 
+
+
+infix 4 ===
+
+-- | Checks for exact exuality between two serialized values
+-- Equality on value is very memory intensive as it's defined on associative
+-- lists and Map equality is implemented. Instead we can be more strict and
+-- require EXACTLY the same value and compare using the serialised bytes.
+(===) :: Value -> Value -> Bool
+(===) val val' =
+  serialiseData (toBuiltinData val) == serialiseData (toBuiltinData val')
+{-# INLINEABLE (===) #-}